requirements.txt
fastapi
Pillow
numpy
starlette
uvicorn
uvicorn[standard]
fastapi[standard]
surya-ocr
torch==2.4.1
torchvision==0.19.1 
torchaudio==2.4.1
psutil
pyjwt
python-dotenv
supabase
pydantic[email]
fastapi-utils
python-multipart
together